Move i22 code into new package structure, adjust signatures and add annotations/docstrings
