+++ pyproject.toml
@@ -15,6 +15,10 @@ classifiers = [
 ]
 description = "One line description of your module"
 dependencies = [
+    "bluesky",
+    "dls_dodal @ git+https://github.com/DiamondLightSource/dodal.git@directory_provider",
+    "dls_bluesky_core @ git+https://github.com/DiamondLightSource/dls-bluesky-core.git@i22_functions",
+    "ophyd_async @ git+https://github.com/bluesky/ophyd-async.git@add-detector-logic",
     "typing-extensions;python_version<'3.8'",
 ] # Add project dependencies here, e.g. ["click", "numpy"]
 dynamic = ["version"]
